+++ packages/twenty-front/src/modules/command-menu-item/display/components/HeadlessCommandMenuItem.tsx
@@ -1,22 +1,35 @@
-import { CommandConfigContext } from '@/command-menu-item/contexts/CommandConfigContext';
-import { useCloseCommandMenu } from '@/command-menu-item/hooks/useCloseCommandMenu';
-import { commandMenuItemProgressFamilyState } from '@/command-menu-item/states/commandMenuItemProgressFamilyState';
-import { useAtomFamilyStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilyStateValue';
 import { useContext } from 'react';
 
+import { CommandConfigContext } from '@/command-menu-item/contexts/CommandConfigContext';
+import { useMountCommand } from '@/command-menu-item/engine-command/hooks/useMountCommand';
+import { isEngineCommandMountedFamilySelector } from '@/command-menu-item/engine-command/selectors/isEngineCommandMountedFamilySelector';
+import { useCloseCommandMenu } from '@/command-menu-item/hooks/useCloseCommandMenu';
+import { commandMenuItemProgressFamilyState } from '@/command-menu-item/states/commandMenuItemProgressFamilyState';
+import { ContextStoreComponentInstanceContext } from '@/context-store/states/contexts/ContextStoreComponentInstanceContext';
+import { useAvailableComponentInstanceIdOrThrow } from '@/ui/utilities/state/component-state/hooks/useAvailableComponentInstanceIdOrThrow';
+import { useAtomFamilySelectorValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilySelectorValue';
+import { useAtomFamilyStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomFamilyStateValue';
 import { isDefined } from 'twenty-shared/utils';
+import { type CommandMenuItemFieldsFragment } from '~/generated-metadata/graphql';
+
 import { CommandMenuItemDisplay } from './CommandMenuItemDisplay';
 
 export const HeadlessCommandMenuItem = ({
-  isMounted,
-  commandMenuItemId,
-  onClick,
+  item,
 }: {
-  isMounted: boolean;
-  commandMenuItemId: string;
-  onClick: () => void;
+  item: CommandMenuItemFieldsFragment;
 }) => {
   const commandMenuItemConfig = useContext(CommandConfigContext);
+  const mountCommand = useMountCommand();
+
+  const contextStoreInstanceId = useAvailableComponentInstanceIdOrThrow(
+    ContextStoreComponentInstanceContext,
+  );
+
+  const isMounted = useAtomFamilySelectorValue(
+    isEngineCommandMountedFamilySelector,
+    item.id,
+  );
 
   const { closeCommandMenu } = useCloseCommandMenu({
     closeSidePanelOnShowPageOptionsExecution: false,
@@ -25,20 +38,29 @@ export const HeadlessCommandMenuItem = ({
 
   const commandMenuItemProgress = useAtomFamilyStateValue(
     commandMenuItemProgressFamilyState,
-    commandMenuItemId,
+    item.id,
   );
 
   if (!isDefined(commandMenuItemConfig)) {
     return null;
   }
 
-  const handleClick = () => {
+  const handleClick = async () => {
     if (isMounted) {
       return;
     }
 
     closeCommandMenu();
-    onClick();
+
+    await mountCommand({
+      engineCommandId: item.id,
+      contextStoreInstanceId,
+      engineComponentKey: item.engineComponentKey,
+      frontComponentId: item.frontComponentId ?? undefined,
+      workflowVersionId: item.workflowVersionId ?? undefined,
+      availabilityType: item.availabilityType,
+      availabilityObjectMetadataId: item.availabilityObjectMetadataId,
+    });
   };
 
   return (
