+++ packages/twenty-front/src/modules/command-menu-item/engine-command/hooks/useMountCommand.ts
@@ -1,147 +1,60 @@
 import { useCallback } from 'react';
 
-import { currentWorkspaceMemberState } from '@/auth/states/currentWorkspaceMemberState';
+import { useBuildTriggerWorkflowVersionCommandState } from '@/command-menu-item/engine-command/hooks/useBuildTriggerWorkflowVersionCommandState';
 import { mountedCommandsState } from '@/command-menu-item/engine-command/states/mountedEngineCommandsState';
-import { type MountedCommandState } from '@/command-menu-item/engine-command/types/MountedCommandState';
-import { contextStoreAnyFieldFilterValueComponentState } from '@/context-store/states/contextStoreAnyFieldFilterValueComponentState';
-import { contextStoreCurrentObjectMetadataItemIdComponentState } from '@/context-store/states/contextStoreCurrentObjectMetadataItemIdComponentState';
-import { contextStoreCurrentViewIdComponentState } from '@/context-store/states/contextStoreCurrentViewIdComponentState';
-import { contextStoreFilterGroupsComponentState } from '@/context-store/states/contextStoreFilterGroupsComponentState';
-import { contextStoreFiltersComponentState } from '@/context-store/states/contextStoreFiltersComponentState';
-import { contextStoreTargetedRecordsRuleComponentState } from '@/context-store/states/contextStoreTargetedRecordsRuleComponentState';
-import { computeContextStoreFilters } from '@/context-store/utils/computeContextStoreFilters';
-import { objectMetadataItemsSelector } from '@/object-metadata/states/objectMetadataItemsSelector';
-import { recordStoreFamilyState } from '@/object-record/record-store/states/recordStoreFamilyState';
-import { getRecordIndexIdFromObjectNamePluralAndViewId } from '@/object-record/utils/getRecordIndexIdFromObjectNamePluralAndViewId';
+import { buildMountedCommandBaseState } from '@/command-menu-item/engine-command/utils/buildMountedCommandBaseState';
 import { useStore } from 'jotai';
 import { isDefined } from 'twenty-shared/utils';
-import { type EngineComponentKey } from '~/generated-metadata/graphql';
+import {
+  type CommandMenuItemAvailabilityType,
+  type EngineComponentKey,
+} from '~/generated-metadata/graphql';
+
+type MountCommandParams = {
+  engineCommandId: string;
+  contextStoreInstanceId: string;
+  engineComponentKey: EngineComponentKey;
+  frontComponentId?: string;
+  workflowVersionId?: string;
+  availabilityType?: CommandMenuItemAvailabilityType;
+  availabilityObjectMetadataId?: string | null;
+};
 
 export const useMountCommand = () => {
   const store = useStore();
 
+  const { buildTriggerWorkflowVersionCommandState } =
+    useBuildTriggerWorkflowVersionCommandState();
+
   const mountCommand = useCallback(
-    ({
+    async ({
       engineCommandId,
       contextStoreInstanceId,
       engineComponentKey,
       frontComponentId,
-      workflowId,
       workflowVersionId,
-      payloads,
-    }: {
-      engineCommandId: string;
-      contextStoreInstanceId: string;
-      engineComponentKey: EngineComponentKey;
-      frontComponentId?: string;
-      workflowId?: string;
-      workflowVersionId?: string;
-      payloads?: Record<string, any>[];
-    }) => {
-      const objectMetadataItemId = store.get(
-        contextStoreCurrentObjectMetadataItemIdComponentState.atomFamily({
-          instanceId: contextStoreInstanceId,
-        }),
-      );
-
-      const objectMetadataItems = store.get(objectMetadataItemsSelector.atom);
-      const objectMetadataItem = objectMetadataItems.find(
-        (item) => item.id === objectMetadataItemId,
-      );
-
-      const currentViewId = store.get(
-        contextStoreCurrentViewIdComponentState.atomFamily({
-          instanceId: contextStoreInstanceId,
-        }),
-      );
-
-      const targetedRecordsRule = store.get(
-        contextStoreTargetedRecordsRuleComponentState.atomFamily({
-          instanceId: contextStoreInstanceId,
-        }),
-      );
-
-      const selectedRecords = (
-        targetedRecordsRule.mode === 'selection'
-          ? targetedRecordsRule.selectedRecordIds
-          : []
-      )
-        .map((id) => store.get(recordStoreFamilyState.atomFamily(id)))
-        .filter(isDefined);
-
-      const filters = store.get(
-        contextStoreFiltersComponentState.atomFamily({
-          instanceId: contextStoreInstanceId,
-        }),
-      );
-
-      const filterGroups = store.get(
-        contextStoreFilterGroupsComponentState.atomFamily({
-          instanceId: contextStoreInstanceId,
-        }),
-      );
-
-      const anyFieldFilterValue = store.get(
-        contextStoreAnyFieldFilterValueComponentState.atomFamily({
-          instanceId: contextStoreInstanceId,
-        }),
-      );
-
-      const currentWorkspaceMember = store.get(
-        currentWorkspaceMemberState.atom,
-      );
-      const systemTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-      const userTimezone =
-        currentWorkspaceMember?.timeZone !== 'system'
-          ? (currentWorkspaceMember?.timeZone ?? systemTimeZone)
-          : systemTimeZone;
-
-      const graphqlFilter = isDefined(objectMetadataItem)
-        ? computeContextStoreFilters({
-            contextStoreTargetedRecordsRule: targetedRecordsRule,
-            contextStoreFilters: filters,
-            contextStoreFilterGroups: filterGroups,
-            objectMetadataItem,
-            filterValueDependencies: {
-              currentWorkspaceMemberId: currentWorkspaceMember?.id,
-              timeZone: userTimezone,
-            },
-            contextStoreAnyFieldFilterValue: anyFieldFilterValue,
-          })
-        : null;
-
-      const recordIndexId =
-        objectMetadataItem && currentViewId
-          ? getRecordIndexIdFromObjectNamePluralAndViewId(
-              objectMetadataItem.namePlural,
-              currentViewId,
-            )
-          : null;
-
-      const baseState = {
-        engineComponentKey,
+      availabilityType,
+      availabilityObjectMetadataId,
+    }: MountCommandParams) => {
+      const baseState = buildMountedCommandBaseState({
+        store,
         contextStoreInstanceId,
-        objectMetadataItem: objectMetadataItem ?? null,
-        currentViewId,
-        recordIndexId,
-        targetedRecordsRule,
-        selectedRecords,
-        graphqlFilter,
-      };
+        engineComponentKey,
+      });
 
-      let commandState: MountedCommandState;
+      const commandState = isDefined(frontComponentId)
+        ? { ...baseState, frontComponentId }
+        : isDefined(workflowVersionId) && isDefined(availabilityType)
+          ? await buildTriggerWorkflowVersionCommandState({
+              baseState,
+              workflowVersionId,
+              availabilityType,
+              availabilityObjectMetadataId,
+            })
+          : baseState;
 
-      if (isDefined(frontComponentId)) {
-        commandState = { ...baseState, frontComponentId };
-      } else if (isDefined(workflowId) && isDefined(workflowVersionId)) {
-        commandState = {
-          ...baseState,
-          workflowId,
-          workflowVersionId,
-          payloads: payloads ?? [],
-        };
-      } else {
-        commandState = baseState;
+      if (!isDefined(commandState)) {
+        return;
       }
 
       store.set(mountedCommandsState.atom, (previousMap) => {
@@ -152,7 +65,7 @@ export const useMountCommand = () => {
         return newMap;
       });
     },
-    [store],
+    [store, buildTriggerWorkflowVersionCommandState],
   );
 
   return mountCommand;
