+++ packages/twenty-front/src/modules/command-menu-item/engine-command/types/MountedCommandState.ts
@@ -18,15 +18,17 @@ export type MountedCommandBaseState = {
   graphqlFilter: Nullable<RecordGqlOperationFilter>;
 };
 
-export type MountedHeadlessFrontComponentCommandState = MountedCommandBaseState & {
-  frontComponentId: string;
-};
+export type MountedHeadlessFrontComponentCommandState =
+  MountedCommandBaseState & {
+    frontComponentId: string;
+  };
 
-export type MountedTriggerWorkflowVersionCommandState = MountedCommandBaseState & {
-  workflowId: string;
-  workflowVersionId: string;
-  payloads: Record<string, any>[];
-};
+export type MountedTriggerWorkflowVersionCommandState =
+  MountedCommandBaseState & {
+    workflowId: string;
+    workflowVersionId: string;
+    payloads: Record<string, any>[];
+  };
 
 export type MountedCommandState =
   | MountedCommandBaseState
