+++ packages/twenty-front/src/modules/command-menu-item/engine-command/utils/isMountedTriggerWorkflowVersionCommandState.ts
@@ -5,5 +5,4 @@ import {
 
 export const isMountedTriggerWorkflowVersionCommandState = (
   state: MountedCommandState,
-): state is MountedTriggerWorkflowVersionCommandState =>
-  'workflowId' in state;
+): state is MountedTriggerWorkflowVersionCommandState => 'workflowId' in state;
